finished tearing apart wooly class.  Lots of things in BallManipulator not compliing due to all the toggle code
