+++ src/org/dirtymechanics/frc/component/arm/grabber/Grabber.java
@@ -6,12 +6,14 @@
 
 package org.dirtymechanics.frc.component.arm.grabber;
 
+import org.dirtymechanics.frc.util.Updatable;
+
 /**
  * Represents a grabber (i.e. the side arms on the robot).
  * 
  * @author Zach Sussman
  */
-public interface Grabber {
+public interface Grabber extends Updatable {
     
     public abstract void openSmall();
     public abstract void closeSmall();
